Vertical standings/results display for teleost
When all the games in a round are completed, teleost, by default,
switches to the standings/results display which shows the standings
table taking up most of the screen and the per-table results on the
bottom.

However, if there are tables with more than three games on them, this
isn't very useful because the results window of this view only shows
three games.

So if there is at least one table in this round with more than three
games on it, use an alternative view - one which shows the standings
table on the left half of the screen, and the results for each table on
the right half of the screen. This way we can fit more games onto the
screen at once, at the expense of horizontal space. We'll automatically
abbreviate people's names to make this a bit more bearable to look at -
we use the player's first name if that's unique, or the player's first
name and initial if that's unique, or their full name if, for example,
there are two "James R"s.