SwissN: fix how table penalty is calculated
The penalty associated with a table of players should be the mean of
the penalties associated with each pair of players on that table,
rather than the highest penalty between a pair on that table, or the
sum of the penalties.

If we use the highest penalty, which we did until this commit, then a
table with a rematch but closely-performing players might not be seen as
more preferable than a table with a rematch but a mixture of stronger
and weaker players.

If we use the sum of penalties, which we did until some time ago, then
tables with more players on have disproportionately higher penalties
than tables with fewer players, which is relevant now we can have tables
of five coexisting with tables of three.