Add option to tell SwissN to start by allowing N rematches
If you have 23 players in the 5&3 table configuration, this is four
tables of five and one table of three. After one round, the software
doesn't realise that it's impossible to find a configuration which
avoids all rematches, so it spends ages looking for configurations where
num_rematches is 0. This makes it go well over the time allowed, and
when it does eventually find a valid configuration after it sets
max_rematches to 1, it takes the first one it finds which is suboptimal.

The SwissN fixture generator now has a control on the form to say "start
by allowing pairings between players who have played each other no more
than <X> times before", where X defaults to 0. In round 2 of the above
scenario, if you start it off at 1 it finds valid configurations much
quicker, and has quite a few to choose from once the time limit is up.