Identical games in same round are not all counted
If in the same round, on the same table, two players play each other
twice and get the same score, only one of those matches gets counted
towards the standings.

This is because the view GAME_DIVIDED does a UNION between two
subqueries which project the round number, table number, player IDs and
scores, and duplicate rows are removed. This needs to be a UNION ALL.

Also, GAME_DIVIDED should really project out the game's sequence number
as well, because (round_no, seq) is guaranteed to be unique. Let's do
that. This actually removes the need to have a UNION ALL, but meh.