Allow the withdrawal and reinstatement of players, and validate player names
The tourney setup page now has controls for withdrawing players, and
(if there are currently withdrawn players) reinstating players. A
withdrawn player is not deleted, and any games they have played or
fixtures they're in will still exist, but the fixture generators will
not include them in any future rounds.

It turns out there was nothing to stop the user from changing a player's
name to a space. Changed the tourney code so that player names which are
empty or consist only of whitespace are invalid.

Also on the subject of validation, if there are no active
(non-withdrawn) players, the fixture generator page now tells you that
you can't generate fixtures rather than offering you a list of fixture
generators, all of which will do something surprising if you tell them
to generate fixtures using an empty list of players.