Treat a draw as half a win
A draw is now worth half a win if the tourney is being ranked on wins
and points. The setup page has a tick box specifying whether to show a
draws column in the standings table, and this is respected by Teleost
and the standings page in atropine. It is intended that the tournament
organiser will tick this box if they intend to allow drawn games rather
than using tiebreaks. Draws are worth half a win even if this box is not
ticked.

Note that a win on a tiebreak (e.g. "55* - 45") counts as a win, not a
draw.