Improve player list interface wrt ratings
When submitting the initial player list, the user must now explicitly
indicate whether they're going to specify ratings themselves, or if they
want ratings to be assigned automatically.

If ratings are to be assigned automatically, there are two choices:
assume the player at the top of the list is the strongest player and
assign ratings at regular intervals down the list, or give all players
the same rating and have an unseeded tournament. In either of these
cases, it is an error for the user to specify ratings manually in the
player list, except for giving a player a rating of zero to indicate
they're a patzer.

Also made some cosmetic changes to HTML forms.