Version 0.6.1, minor changes
Allow deletion of knockout rounds. Previously when you tried to delete a
round it would only let you delete the last prelim round.

Introduce a checkbox in the Display Control page to specify whether you
want the Standings/Results page to have the vertical format when in Auto
mode. This means that in Auto mode, when all the games in a round have
been played but the fixtures for the next round have not yet been
generated, it'll display the standings table and last round's results
side by side rather than one atop the other.

Fixture generation: put the name of the fixture generator (e.g. Swiss
Army Blunderbuss, Random Fixture Generator) at the top of the page when
it asks you for any information required for fixture generation.