Don't let user add player who already exists
If you use the new "add player" feature to add a player mid-tournament,
and there's already a player by that name, you get an ugly Python stack
trace because the attempt to insert into the PLAYER table violates its
uniqueness constraint on NAME, which raises an SQLite exception.

We should catch this case before we run the SQL to insert a player,
so we can present an exception to the user in the normal way with a
more helpful message.