Make division setup cope better with withdrawn players
The division setup page can split your field of players up into N
divisions. Before, it didn't treat withdrawn players specially, but now
it does. First it divides the active players into divisions so that the
number of active players in each division satisfies the "each division
must have a multiple of X players" constraint. Then it adds the
withdrawn players to divisions appropriate for their rating. This may
mean a division has, say, 17 players in it when you asked for it to be
an even number, but the number of *active* players will be even.

Changed some labels to give a bit more guidance at the start of the
tourney, like "no, you can't set divisions yet, you haven't given me a
player list" and "if you submit this new player list, any existing
division setup will be lost".

Some cosmetic changes, and a fix to a bug in the swissN generator caused
by a typo, which caused an exception to be raised if you tried to
generate fixtures in a 5&3 configuration.