Enable tournament ratings, do other fixes
Tournament ratings are now visible to the user in the standings table
(if ticked in the setup page) and in a player's profile. The constants
are tweakable on the setup page in the tourney rules section.

Important columns in the standings table (that is, columns that are used
to rank players in order) are now bold.

Links to export results and standings as HTML or plain text are now in
the sidebar rather than on the tourney setup page.

The tourney setup page has a few more cosmetic improvements and
explanations.