Make manual fixgen save incomplete form, warn when losing unsaved work
The manual fixture generator page which gives you drop-down boxes to
select which players are in which groups now warns you if you try to
navigate away from the page before pressing Submit. If you do press
Submit when the form is incomplete, the settings are saved, and next
time you open the manual fixture generator, if there is an existing set
of settings saved, it asks you if you want to pick up from where you
left off.