Start of 0.7.1, various minor fixes
Fix the CoEdi generator, even though nobody uses it any more. It's now
division-aware and doesn't break when confronted with withdrawn players.

If a player has just been added to the tournament, they don't have a
defined percentage of games they've played first in, which the Swiss
generator uses. Previously, for games involving new players the order
would be random; now the new player is treated as having a played-first
ratio of 50%.

In Teleost, you can now press space to force the screen to move to the
next page.

Teleost and the standings page now show withdrawn players in grey.

The general setup page has a bit more explanation of what the "show
draws column" option does.