Round Robin fixture generator, and per-division fixture generation
The Round Robin fixture generator has been added. This generates, for
each division of N players, N-1 rounds at once, and every player in a
division plays every other player in that division once.

Fixture generation itself has undergone a change to allow you to
generate fixtures for selected divisions rather than all divisions. So
if you have four divisions A-D of eight players and one division E of
six players, and you generate a round robin, divisions A-D will have
games for 7 rounds but E will only have games for 5. After round 5, you
might want to generate games for division E for round 6 (and later 7)
using the Swiss generator, for example.

So now, when you ask to generate fixtures and select your fixture
generator, it asks you which divisions you want to generate for, and
from which round number. The round number for a division defaults to the
round number after the highest round which contains games in that
division.

The knockout fixture generator has been disabled by renaming it. Too
much of a pain in the arse to maintain for something that's never used.

Teleost now copes with there being future rounds in the database that
aren't the current round. It will now focus on only the round currently
being played, or the next round if the current round has finished.

The Teleost background image, and the web favicon, is now a round robin.