HTML and CSS changes
When you export the results to HTML, the CSS is now included in the
<style> tag in the HTML itself, rather than referenced in a <link> tag,
so you can take the HTML file and send it to someone else without losing
the style information.

Warning boxes are now a thing - they look like the warning sign for
ducks in the road.

The fixture generators menu is now a tiny bit less ugly.