0.7.3: Fix divisioned fixture generation, and other improvements
Divisioned fixture generation was completely broken in 0.7.2. It would
only generate fixtures for division A, and when I fixed that it would
put matches on the same table that didn't belong there. Fixed.

You can now specify that certain matches are not to be counted in the
standings, such as finals. These have a game type of "N" rather than
"P". You can specify this at fixture generation time, or after the
fixtures are generated by fiddling with the checkboxes on the Edit
Fixtures page.

Divisions: it's now easier to specify which players are in which
division, before the tournament has started. In the big text box on the
general setup page where you enter the player names, if you enter the
player name "-" it means the players after that point are in a new
division.

In the division setup page you can rename the divisions from "Division
A", "Division B", etc to whatever you want, e.g. "Pros" and "Noobs". You
can still assign players to divisions using this page, but now you can
do it by standings position as well as by rating.

Random results button (set showrandomresultsbutton=1 in the URL of the
score editor page): this would try to divide by zero if there are two
players with rating 0 playing each other. This is now fixed - such games
are "randomly" given the score 0-0.