Fix double-counting table nos when generating games for non-empty round
If (say) round 1 already has some fixtures generated for it in Division
A, using tables 1 to 4, when you generate some fixtures for Division B,
you'd expect the fixtures to start from table 5. When you do this,
you'll find they start from table 9.

This is because the code that's supposed to avoid duplicating table
numbers exists twice.