Show any rematches in fixture generator's output
When the fixture generator generates fixtures, fixturegen.py shows us
what fixtures it's generated. It's not obvious from this page whether
it's generated any rematches.

So if a match is a rematch, put a suitable message in the remarks column
on the right.

Also, the player names in the fixture generator's output are now links
to the player's page, which open in a new window.