Improvements to drop-down boxes in manual fixture generator
In the manual fixture generator, after you "Submit" a
partially-completed fixture list, the drop-down boxes for choosing
players will only contain the players yet to be assigned a table.
Assuming you Submit regularly when entering the fixtures, this makes
things a lot easier for the organiser when entering the last few tables.
This behaviour can be switched off with a checkbox.