Add Tuff Luck page to atropine
Teleost has a Tuff Luck display, so it's sensible for atropine to be able
to show the operator the same information.

A player's Tuff Luck (or Tuffness?) is the sum of the losing margins of
their three closest losing games. The player who lost three games by the
lowest aggregate margin wins the Tuff Luck prize.