Add Overachievers page to atropine
Same as with Tuff Luck - Teleost shows it but atropine doesn't, so let's
fix that.

The Overachievers page shows, for each division, the list of players and
the number of positions above their seed position they are in the
standings table. A player's seed is calculated from their rating, with the
top-rated player in a division being #1, and so on down.

If the Overachievers page is viewed and all players in a division (bar
0-rated patzers) have the same rating (because you selected the "this
tournament is not seeded" option when you entered the player list), it
gives you a warning that the table doesn't mean much.