Name-to-table index for teleost, and animated page scrolling
Teleost now has a name-to-table index page, which can optionally be
used in Auto mode in place of the fixture display you see after new
fixtures are generated. This should make it easier for players to find
their table - now they just have to find their name in an alphabetical
list rather than read a fixture list sorted by table looking for their
name.

When a Teleost screen scrolls to a new page within the same view, this
scrolling action is now animated. This makes it a bit more obvious that
the page has changed - if you're reading along a row in, say, the
standings table, it's possible not to notice the page has changed
between reading the name column and the wins/points columns. The animated
scrolling can be switched off in the Teleost display page in atropine, if
required.