Fix King of the Hill generator bug with withdrawn players
If a player has withdrawn, and you use the King of the Hill generator,
it sees the withdrawn player in the standings table and can't find them
in the list of active players for that division, so it throws an
exception and you get a message saying it's never heard of the withdrawn
player.

Fix it so that if someone in the standings table doesn't exist in the
list of active players, shrug and carry on.