First videprinter entries should appear at the bottom
If the videprinter has fewer rows on it than there is space for, we
should leave space at the top rather than the bottom, so new entries
always appear at the bottom no matter how much or little space there is
on the screen. If we put the first entry on the top line, the scrolling
animation looks wrong.