Fix fixture generation bug where new round can be half added
If you play round 1, then generate fixtures for round 2, but don't press
Accept yet, then change a player's name, then press Accept, the fixtures
are rejected because they contain a player name it doesn't recognise any
more.

This isn't the bug (although perhaps we should revisit this another
time) - the bug is that when this happens, round 2 still seems to have
been created without any games. As far as the teleost display and the
Games link in the sidebar are concerned, the current round is round 2.
It should still be round 1, because fixture submission for round 2
failed.

Solution: if fixtures couldn't be generated for whatever reason, don't
call the tourney's name_round() method.