Improvements to Teleost
Added a "standings only" view.

Keys F1-F10 now select the various modes, and F12 goes back to auto
mode.

Press P to pause/unpause page scrolling. While it's paused, you see a
little pause icon in the top right corner. You can still force the view
to change to the next page by pressing space.

Press H for the help that now gets shown on startup.

Standings view: When the table scrolls to the next page and that's a
new division, make it so that the division name changes during the
scroll animation rather than after it.