Option to tell the Swiss fixgen not to make a player play Prune
If the tournament director wants to ensure a player cannot play Prune
despite not having played them before (say because the player played
a Prune in yesterday's tournament) they can now tick a box in the
player's configuration page to say "Avoid Prune". This tells the Swiss
fixture generator to behave as if the player has already played a Prune.
Note that it doesn't affect the behaviour of any other fixture
generator.