Get rid of incorrect warning about messing up division setup
If you've set up divisions but not generated any games yet, you see a
warning on the tournament setup page which tells you that if you
resubmit the player list you'll lose the division setup. This appears to
be nonsense.

For some time now, the player list has had dashes in it to mark the
division boundaries. If you leave them in when you edit the player
list, the divisions will still be there.