New results entry interface
The results entry page has now undergone a revamp. When you want to
enter a result, enter the players' names and then the scores, and then
press enter to submit.

The players' names are autocompleted for you, so it's very rare that
you'll have type a player's full name.

Most importantly, you can type the players' names in either order, which
means you can type the names and scores as they appear on the scoresheet
rather than having to mentally reverse the score if the players on the
screen are the other way round from the order on the sheet. Being handed
a scoresheet showing "Alice 65-43 Bob" when the game list says
"Bob  -  Alice" is the number one cause of accidentally reversed
results, and this interface is intended to make that mistake less
likely.

The main results entry page also contains a representation of the
videprinter, with which the user can check previously-entered results.

It also contains the "blinkenlights", which is a diagram showing a
numbered box for each table, each of which contains one little coloured
box for each game - red for unplayed, grey for played. This shows an
at-a-glance representation of which tables have finished, which are
playing slowly, and how many more results there are to come.

You can click on results in the videprinter to load them into the main
results entry form for review or correction. You can also click on games
in the blinkenlights to do the same thing.

Finally, the old results entry interface, which showed all the games in a
list with an edit box for each score, still exists - it's been moved
from games.py to gameslist.py. games.py is now the new interface.

There are some very rare and esoteric things the old interface could do
that the new one couldn't. One is that you could enter a tiebreak score
where the margin isn't 10 (in case the rules at that event are different
for some obscure reason). The other is that if the same two players are
playing each other twice in a round (there used to be fixture generators
that could do this), then you need to use the old interface to enter the
scores, because the pair of player names isn't a sufficiently unique key
to identify which game is having its result submitted.