Configurable tab-order for results entry interface
There is now a "Preferences..." link, which, like the other links in the
sidebar which open in a new window or tab, has an icon next to it
telling the user this.

The plural "Preferences" is somewhat over-egging it, because currently
there is only one preference: the tab-order for the results entry
interface. You can choose from Name-Name-Score-Score (the default),
Name-Score-Name-Score, and Name-Score-Score-Name.

Preferences are stored in a separate database, prefs.db, in atropine's
top-level directory. This means they're not specific to a tournament.
They're intended only for tweaking the behaviour of the user interface,
and not anything to do with tournament rules or results.

Relatedly: if you fill in one name, then the other, then tab out of the
second name field, usually a Javascript event will fill in the score
boxes with the current score for that game. This has changed slightly.
This now only happens if the score boxes are either blank or filled in
with something the script put there rather than the user. This is
necessary, because if your preferred tab-order is Name-Score-Name-Score,
you type "Alice", tab, "44", tab, "Bob", tab - and then it would load
the Alice-Bob game and put the current score (probably blank) into the
score boxes, removing the 44 you typed in.

So, common workflows are:
* User enters the names "Alice" and "Bob". That game hasn't already been
  played. The score boxes are currently blank so they're overwritten by
  the currently-recorded score for that game, which is also blank. User
  types in the score and submits the result.
* User enters the names "Alice" and "Bob", and that game has already
  been played. The recorded score for that game is automatically put
  into the score boxes, inviting the user to edit it.
* User enters the names "Alice" and "Bob", a game which has already been
  played. The recorded score for that game is automatically put into the
  score boxes as before. Then the user realises they typed the wrong
  name - actually they wanted to enter the result between Alice and
  Fred. User changes "Bob" to "Fred", presses tab, and the current
  Alice-Fred score is put into the score boxes in place of the
  Alice-Bob score.
* User's tab order is Name-Score-Name-Score. User enters "Alice", tab,
  "44", tab, "Bob", tab - and because one of the score boxes contains
  data the user typed in, the currently-recorded Alice-Bob score is not
  splatted over the boxes. User types in Bob's score in the last box and
  all is well.

Also fixed a case where you could click a tiebreak game in the
Blinkenlights to load it into the results entry pane, but the tiebreak
checkbox's checked state wouldn't be set.