Always focus score1 when we click a game in videprinter or blinkenlights
When you click a game in the videprinter or on the blinkenlights panel,
the player names, and currently-recorded score (if any) are loaded into
the results entry pane.

If the game hasn't been played yet then we also give focus to the first
score box, in anticipation of the user typing in a score. But if the
game already has a score recorded, we don't do this.

If you click a game that's already got a score associated with it, then
when we load the score into the results entry pane we should focus the
first score box and highlight its text, because the next thing the user
wants to do will probably be to edit the score.