New Teleost implementation
This is the start of a new Teleost, which will eventually remove the
need for pygame. The New Teleost is just a page in the web interface
with lots of Javascript on it. You open that in a new window, drag it
over to the second screen and full-screen it.

Currently the new Teleost screen has only one view: the standings and
videprinter view. More views will be added shortly.