New Teleost: more views, and automatic view switching
Auto, Standings, Standings/Results, Standings/Videprinter,
Fixtures and Table Name Index are now done.

Vertical Standings/Results doesn't exist in New Teleost. It was hardly
ever used, and is a real pain to do in an environment when you can't
squish a text label into a space.

Teleost options are no longer defined in the database, they're
hard-coded in countdowntourney.py. Their values are still stored in
the database.