Enable a game to be recorded as lost by both players
Sometimes, a situation can arise where you want to adjudicate a game
as having been lost by both players. This is most useful if both
players fail to turn up. Until now the best way to deal with that has
been to enter the score as 0-0, but this isn't quite satisfactory
because it gives both players half a win.

You can now enter a score as 0-0* (that is, a 0-0 draw with the
tiebreak flag set) and this will be taken as a scoreless loss for
both players. Neither player is credited with a win or a draw for
that game. The score appears in New Teleost and in various places
in the web interface as "X-X".

The results entry interface is now aware of this - if you type a zero in
both score boxes, the tiebreak tickbox will appear, but with its label
asking if it's a loss for both players rather than if it's a tiebreak.