Put link to open New Teleost in sidebar
The link to open the new Teleost display is now in the sidebar, and even
has a nice blue rounded rectangular background in case it wasn't obvious
enough.

Some tweaks to the display options page to remove references to having
"Teleost running".

The filename of the CGI script to tweak the display options
(webroot/cgi-bin/teleost.py) and the filename of the new display
(webroot/cgi-bin/display.py) are blatantly the wrong way round in terms
of describing what they do. Some sort of correction for this has now
been made: the display options page is now called displayoptions.py.