Highlight players who have definitely reached the final
First, a load of unused cruft about tournament ratings has been moved
from the tourney setup page to a new "advanced setup" page.

In its place are some controls for the organiser to tell atropine how
many games each player is expected to play, how many rounds there are
expected to be, and how many players qualify for the final (or
semi-finals, or quarter-final, or whatever).

If a player is mathematically certain of qualifying, no matter what
future results come in, they're now highlighted in green on the
standings table in Teleost and on the atropine standings page.

When we decide whether someone can be pushed out of the qualification
zone, we take into account the remaining fixtures. For example, suppose
the top two players go through the final of a tourney where each player
plays six games. Near the end of the day, the top of the standings
table looks like this:

              P  W  Pts
1st  Alice    6  6  370
2nd  Bob      5  5  360
3rd  Charlie  5  5  358
4th  Dave     6  4  371
5th  Eve      6  4  356
...etc

It appears that Alice isn't quite certain of qualification yet, as Bob
and Charlie might also get 6/6 and overtake Alice on points, pushing
Alice down to third. However, if Bob and Charlie's last game is
against each other, only one of them can win, Alice would be
guaranteed at least 2nd place, and she'd be highlighted green in the
standings.

The reason we need to know both the number of possible games and the
number of rounds is because if we have a latecomer who only started in
round 3 and won both games, we don't want atropine thinking they've
still got four games to play and might storm ahead of the leaders. If
we're in the last round then atropine ignores the "played" column of
the standings, and assumes that the games generated for this round in
this division will be the last.

Additionally:
The "records" or "notable games" view now exists. Still doesn't have a
suitable icon in the display setup page though.

Fix a mistake in the PagedTableView implementation - the PagedTableView
should take care of remembering which was the last game state revision
seen, rather than having the same code in all the subclasses.