Make max_games_per_table value a per-round value
In jsonreq.py we were returning max_games_per_table, which is the
maximum number of games on one table in all the games in that division.
This is used by the round results view to help determine whether to
group games by table or not (if there's only one game per table it
doesn't bother).

This should instead be per-round. If this is .co.mk, and we're in
round 2, and round 1 had 6 games on each table, and round 2 has 1 game
on each table, the useful answer is 1, not 6.