Fully manual fixture generator
There is now a "fully manual" fixture generator, which can generate
an arbitrary number of fixtures for a round without the requirement that
all active players play. For each applicable division, the user specifies
how many games are to be created, and what their game type is to be.
Then on the next screen the user is shown that many rows of two
drop-down boxes to select the players.

"game types" up until now have been "P" (prelim, counts in the standings
table) and "N" (not a prelim, doesn't count towards the standings).
There are now "QF" (quarter-final), "SF" (semi-final), "3P" (third place
play-off) and "F" (final) as well. This is not a return to the old
"knockout fixture generator" - atropine doesn't know how these game
types relate to each other. All it knows is that there are these
different type codes, each associated with a description.

Round names now work a little differently. If no round name is specified
by the fixture generator when it generates the fixtures (the new fully
manual generator does not specify a round name), then the name is blank
in the ROUNDS table. A blank name in the ROUNDS table indicates that the
round name should be worked out from the types of the games in it. If
all a round's games are of type "QF" then the round is called
"Quarter-finals". If all a round's games are of type "SF" then it's
called "Semi-finals", and likewise for "3P" (third place) and "F"
(final). If none of these conditions are true then the round name
defaults to the word "Round", followed by a space, followed by the round
number.

This means if you generate a quarter-final round, and all the games in
that round are of type "QF", it'll show up as "Quarter-finals" on
Teleost.