Add "rerate players by player ID" button to player setup page
This is for the very specific case that cropped up at the last COLIN. At
the end of the tournament, the Overachievers page didn't show anything
useful because at the start of the tournament, the radio button to say
"the players in this list are in rating order" hadn't been checked.

At the end of the tournament there's no opportunity to go back and
set this, because you can't resubmit the player list once the tournament
has started. The only remedy would have been to set each player's rating
individually and manually. (On this occasion I just made a Python script
to solve the problem.)

The player setup page now has a "rerate players by player ID" button,
which sets the first player's rating to 2000, the next player's rating
slightly lower than that, and so on down to the last player who gets a
rating of 1000. Any player with a rating of 0 is left unchanged.