Disregard non-players when working out who has definitely qualified
When we work out whether someone can't be prevented from finishing in
the qualification zone, we should remove withdrawn players and prunes
from the equation, because they can't qualify.

Also modify the four-to-a-table game order so that the first two games
can be played together, then the middle two, then the last two.