Stop using print() and use html.escape not cgi.escape
print() makes assumptions about the console's preferred character
encoding which may not be correct. On Windows it converts the string
to some old Windows codepage or other, when we want it to output
UTF-8. So all print() statements have been replaced with
cgicommon.writeln(), and cgicommon.writeln() writes UTF-8 encoded data
directly to stdout.

Also replaced all calls to cgi.escape() (now deprecated) with
html.escape(). The only difference seems to be that html.escape() has
the "quote" parameter set to True by default, so it's safer.