Make Round Robin generator work with an odd number of players
There's no reason why we can't generate a Round Robin with an odd
number of players. It just means one player has to sit out each round.

Internally, the Round Robin fixture generator now adds an imaginary
player to the list to make the number of players even, then behaves as
before except that if a player is drawn to play against the imaginary
player than that fixture isn't generated.