Add list of games to new result entry interface
The new result entry interface now has a list of games on the right hand
side, or below the main pane if it won't fit on the side. This can be
useful if the organiser needs to read out the list of fixtures, for
example.

If a game in this list is clicked, it has the same effect as clicking
the relevant button on the Blinkenlights - the player names and score
for that game are loaded into the edit boxes.