"Random from Seeded Pots" fixture generator
This is like the Random fixture generator except that it divides all the
players by rating into N pots (N is the desired number of players on each
table), and each table then consists of one randomly-selected player from
each pot. This is how it's done at MK.

People often wonder, who is "Random" and where is "Seeded Pots"?