Option not to switch to Fixtures display until all divs generated
Suppose we have a CoBrum style event, where the Round Robin fixtures
for rounds 1-7 are generated for the top division in advance, and the
random or Swiss fixtures are generated round-by-round for the open
division.

After we finish round 1, the Auto display mode thinks that the fixtures
for round 2 have already been generated, because there are fixtures for
the top division in that round, so it immediately switches to the
Fixtures display.

What we usually want in this situation, however, is for it to stay on
the Standings/Results display (as if round 1 has finished but round 2
hasn't been generated yet) until we generate the round 2 fixtures for
the open division.

So there's now a checkbox, which is ticked by default, to tell the
Auto display mode only to consider a round to be "generated" if games
are generated in that round for *every* division.