Impose time limit on calculation of guaranteed qualifiers
If we generate N-1 of the N rounds, and don't play them, the code in
qualification.py can take forever to work anything out, because all
the players are in joint first place. This affects anything that
generates a standings table.

Now, if this takes more than five seconds, it will throw an exception
recommending the user to disable qualification calculation by setting
the number of qualification places to zero.