Introduce "accessible tables" feature
The tournament director can now specify that a certain set of tables
are "accessible" (General Setup page), and that certain players require
accessible tables (Player page). This does not affect who plays whom -
it only affects which groups go on which table numbers.

This has necessitated a new database column (require_accessible_table
in the PLAYER table), and a new database table to store information
about tables (called BOARD, because we can't call it TABLE), so the
version number is bumped to 1.0.4.

Calls to methods in countdowntourney.py which ask about accessible
tables for tourney databases generated before 1.0.4 will return canned
replies implying there aren't any accessible tables or players who
require them.

Fixture generators now return a GeneratedGroups object (defined in
fixgen.py), rather than a messy dictionary of rounds and fixtures.
Fixture generators add groups of players (lists of Player objects)
to the GeneratedGroups object, associating with each group a round
number, division number and other attributes.

When the fixture generator returns details of what the groups are,
it's up to the code in py/countdowntourney.py to generate fixtures
from those groups and decide which groups are going on which tables.

The "accessible groups" - that is, groups which contain at least one
player who requires an accessible table - are placed first, on
accessible tables if available. They are placed in alphabetical order
of the first-alphabetical player in each group, which means that
where possible, a player who requires an accessible table should get
put on the same table each round.

Then the other groups are placed.

If you generate fixtures for (say) Division A, and table 1 is the
accessible table, then if there are no players in Division A who
require accessible tables, but there is such a player in Division B
whose fixtures have yet to be generated, the table numbers for
Division A will deliberately skip over table 1 so that Division B
can use it.

It is possible for countdowntourney.make_fixtures_from_groups() to
give an accessible table to a group who doesn't need it, but it only
does that if there would still be enough accessible tables for
players who need them.