Revamp Players page
The Players page (player.py) should now be a bit easier to use. The
player list now appears constantly in a pane on the left, and clicking
on a player's name causes that player's details to appear in the pane on
the right, where they can be edited.

This means when you've edited and changed a player's details, you don't
have to click back to the "main" Players page to get to another player's
details - the player list is still there so you can just pick whichever
player you want to see next from there.

Players in the player list may also have symbols alongside their name.
These include a wheelchair symbol (player requires accessible table) a
hollow circle (player is a Prune) and a black circle (player will be
considered to have already played a Prune). If a player is withdrawn,
their name is greyed out to indicate this, but it can still be selected
in order to unwithdraw them.

Above the list of players is an "Add new player..." link, which puts an
"Add New Player" form in the pane on the right.

The "I'm a muppet who forgot to tell you I was ordering players by
rating when I gave you the player list" button is now on the
Overachievers page rather than the Player page. The Overachievers page
won't give you an overachievers table if all the non-pruney players have
the same rating.